Remove Manila V1 API client
The Manila V1 API has been deprecated since 2016 and tests
stopped using it in change Id90044bca44387864aee8984bc2366c089c87cfc.
Complete this removal by:

1. Making SharesV2Client inherit directly from RestClient
   instead of the V1 SharesClient
2. Add microversion support added to all methods
3. Removing V1 client registration from plugin.py
